- ajout du systeme d'ordre pour les batiment de stockage (Accept / Reject / Empty), Obtain n'a pas encore été implémenté.

+++ src/com/exiro/constructionList/Road.java
@@ -18,6 +18,7 @@ public class Road extends Construction {
     Agora agora;
 
     boolean start = false;
+    boolean inDestructible = false;
 
     public Road(boolean isActive, ObjectType type, ArrayList<Case> cases, int cost, int deleteCost, int xPos, int yPos, int xLenght, int yLenght, float cachet, City city, boolean built, boolean isFloor) {
         super(isActive, type, cases, cost, deleteCost, xPos, yPos, xLenght, yLenght, cachet, city, built, isFloor);
@@ -134,7 +135,7 @@ public class Road extends Construction {
 
     @Override
     public void delete() {
-        if (!this.start && agora == null) {
+        if (!this.start && agora == null && !this.inDestructible) {
             super.delete();
             city.getPathManager().deleteRoad(this);
         }
@@ -154,6 +155,14 @@ public class Road extends Construction {
         TileImage img = ImageLoader.getImage("Zeus_General", 3, 12 + random.nextInt(3));
         assert img != null;
         setImg(img);
+    }
+
+    public boolean isInDestructible() {
+        return inDestructible;
+    }
+
+    public void setInDestructible(boolean inDestructible) {
+        this.inDestructible = inDestructible;
     }
 }
