+++ src/com/exiro/render/EntityRender.java
@@ -102,7 +102,8 @@ public class EntityRender {
                     if (startY < p.getY()) {
                         for (int j = startY; j < p.getY(); j += defaultObject.getSize()) {
                             Case c = map.getCase(i, j);
-
+                            if(c == null)
+                                continue;
                             ArrayList<Case> place;
                             place = defaultObject.getPlace(i, j, defaultObject.getSize(), defaultObject.getSize(), GameManager.currentCity);
                             if (place.size() == defaultObject.getSize() * defaultObject.getSize()) {
@@ -282,7 +283,7 @@ public class EntityRender {
         for (int i = (int) Math.min(startX, currCase.getxPos()); i < Math.max(startX, currCase.getxPos()); i += defaultObject.getSize()) {
             for (int j = Math.min(startY, currCase.getyPos()); j < Math.max(startY, currCase.getyPos()); j += defaultObject.getSize()) {
                 Case c = GameManager.currentCity.getMap().getCase(i, j);
-                if (c.getObject() != null) {
+                if (c != null && c.getObject() != null) {
                     c.getObject().delete();
                 }
             }
